requirements.txt
pyrogram>=2.0.0
selenium>=4.0.0
vk-api>=11.9.0
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
fake-useragent>=1.1.0
undetected-chromedriver>=3.4.0
python-dotenv>=0.19.0
tgcrypto>=1.2.0